app/requirements.txt
Django==4.2.9
pytz==2022.4
tzdata==2022.4
django-admin-interface==0.22.0
python-telegram-bot==13.10
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.4
requests==2.32.3
python-telegram-bot==13.10
